+++ lib/config/home.nix
@@ -4,6 +4,8 @@
   hydenix.hm = {
     enable = true;
 
+    editors.antigravity = true;
+
     theme = {
       enable = true;
       active = "BlueSky";